Fix problems with parameters assignment
When ipcRenderer listener is created, there would be an error that says
that argument of type `T` is not assignable to a parameter of type `T |
undefined`. This happened because the data parameter is optional in
answerMain callback. TypeScript automatically adds possible `undefined`
type when a parameter is optional in a function so this is now fixed.