Upgrade Electron, XO and tsd
Upgrade Electron to v12 which would require the move to
@electron/remote. Minor changes were made in the code since there is
only one place where it is really used. XO and tsd are upgraded to the
latest version. There where some errors in the ts declaration file
tests because of that so there were some minor changes mate to them
but functionally nothing is changed.